@@ -4,23 +4,23 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "framer-motion": "^4.1.17",
-    "gh-pages": "^3.2.3",
+    "@testing-library/react": "^15.0.7",
+    "@testing-library/user-event": "^14.5.2",
+    "framer-motion": "^11.2.9",
+    "gh-pages": "^6.1.1",
     "json-loader": "^0.5.7",
-    "mongoose": "^5.12.14",
-    "react": "^16.0.2",
+    "mongoose": "^8.4.0",
+    "react": "^18.3.1",
     "react-animations": "^1.0.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.3.1",
     "react-fade-in": "^2.0.1",
-    "react-pdf": "^5.3.2",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "react-slideshow-image": "^3.5.0",
-    "react-transition-group": "^4.4.2",
+    "react-pdf": "^9.0.0",
+    "react-router-dom": "^6.23.1",
+    "react-scripts": "5.0.1",
+    "react-slideshow-image": "^4.3.1",
+    "react-transition-group": "^4.4.5",
     "react-typist": "^2.0.5",
-    "web-vitals": "^1.1.2"
+    "web-vitals": "^4.0.1"
   },
   "scripts": {
     "start": "react-scripts start",
