@@ -8,7 +8,7 @@ import "./HomeSkills.css";
 export default function HomeSkills() {
     const [buttons, setButtons] = useState([
         { name: 'Languages', selected: true},
-        { name: 'Tools', selected: false},
+        { name: 'Engineering', selected: false},
         { name: 'Office', selected: false},
         { name: 'Design', selected: false},
     ])
@@ -30,13 +30,13 @@ export default function HomeSkills() {
             
             <div className="buttonContainer row mobile-hide">
                 <a href="#aleyda" onClick={handleClick} className={`col-lg-2 col-md-12 btn btn-danger  ${buttons[0].selected ? 'selected' : ''}`}>Languages</a>
-                <a href="#aleyda" onClick={handleClick} className={`col-lg-2 col-md-12 btn btn-primary ${buttons[1].selected ? 'selected' : ''}`}>Tools</a>
+                <a href="#aleyda" onClick={handleClick} className={`col-lg-2 col-md-12 btn btn-primary ${buttons[1].selected ? 'selected' : ''}`}>Engineering</a>
                 <a href="#aleyda" onClick={handleClick} className={`col-lg-2 col-md-12 btn btn-warning ${buttons[2].selected ? 'selected' : ''}`}>Office</a>
                 <a href="#aleyda" onClick={handleClick} className={`col-lg-2 col-md-12 btn btn-success ${buttons[3].selected ? 'selected' : ''}`}>Design</a>
             </div>
             <div className="buttonContainer row mobile-unhide">
                 <div onClick={handleClick} className={`col-lg-2 col-md-12 btn btn-danger  ${buttons[0].selected ? 'selected' : ''}`}>Languages</div>
-                <div onClick={handleClick} className={`col-lg-2 col-md-12 btn btn-primary ${buttons[1].selected ? 'selected' : ''}`}>Tools</div>
+                <div onClick={handleClick} className={`col-lg-2 col-md-12 btn btn-primary ${buttons[1].selected ? 'selected' : ''}`}>Engineering</div>
                 <div onClick={handleClick} className={`col-lg-2 col-md-12 btn btn-warning ${buttons[2].selected ? 'selected' : ''}`}>Office</div>
                 <div onClick={handleClick} className={`col-lg-2 col-md-12 btn btn-success ${buttons[3].selected ? 'selected' : ''}`}>Design</div>
             </div>
