@@ -1,28 +1,31 @@
 export const education = [
     {
-        logo: "https://i.imgur.com/NZSGKfu.png",
-        name: "University of California, Irvine",
-        shortName: "UC Irvine",
-        time: "2012-2017",
-        graduation: "2017",
+        logo: "https://i.imgur.com/ic6MzXS.png",
+        name: "California State University, Fullerton",
+        shortName: "Cal State Fullerton",
+        time: "2022-2025",
+        graduation: "2025",
         type: "Degree",
-        degree: "Bachelor of Science",
-        major: "Applied Mathematics",
-        minor: "Environmental Science",
+        degree: "Master of Science",
+        major: "Computer Science",
+        minor: "",
         coursework: [
-            "Abstract Algebra",
-            "Abstract/Discrete Mathematics",
-            "Data Modeling", 
-            "Differential Equations", 
-            "Elementary Analysis", 
-            "Group Theory", 
-            "Linear Algebra", 
-            "Multi-Variable Calculus", 
-            "Numerical Analysis", 
-            "Partial Differential Equations", 
-            "Probability", 
-            "Proofs and Theory"]
-    }, 
+            "Advanced Database Management", 
+            "Algorithm Engineering",
+            "Assembly Language",
+            "Backend Engineering",
+            "C++", 
+            "Compilers",
+            "Data Structures", 
+            "Network Security",
+            "Operating Systems",
+            "OOP", 
+            "Python", 
+            "Software Engineering",
+            "Software Management",
+            "Software Validation / Verification",
+        ]
+    },
     {
         logo: "https://i.imgur.com/J1kkzc0.png",
         name: "General Assembly",
@@ -48,28 +51,32 @@ export const education = [
             "React", 
         ]
     },
-
     {
-        logo: "https://i.imgur.com/ic6MzXS.png",
-        name: "California State University, Fullerton",
-        shortName: "Cal State Fullerton",
-        time: "2022-2025",
-        graduation: "2025",
+        logo: "https://i.imgur.com/NZSGKfu.png",
+        name: "University of California, Irvine",
+        shortName: "UC Irvine",
+        time: "2012-2017",
+        graduation: "2017",
         type: "Degree",
-        degree: "Master of Science",
-        major: "Computer Science",
-        minor: "",
+        degree: "Bachelor of Science",
+        major: "Applied Mathematics",
+        minor: "Environmental Science",
         coursework: [
-            "Algorithm Engineering",
-            "Assembly Language",
-            "C++", 
-            "Compilers",
-            "Data Structures", 
-            "Database Management", 
-            "Operating Systems",
-            "OOP", 
-            "Python", 
-            "Software Engineering",
-        ]
-    }
+            "Abstract Algebra",
+            "Abstract/Discrete Mathematics",
+            "Data Modeling", 
+            "Differential Equations", 
+            "Discrete Mathematics",
+            "Elementary Analysis", 
+            "Group Theory", 
+            "Linear Algebra", 
+            "Multi-Variable Calculus", 
+            "Numerical Analysis", 
+            "Partial Differential Equations", 
+            "Probability", 
+            "Proofs and Theory"]
+    }, 
+    
+
+    
 ]
