@@ -1,5 +1,53 @@
 export const workexperience = [
     {
+        logo:       "https://i.imgur.com/hqiI1Qy.png",
+        name:       "Paramount+",
+        shortname:  "Paramount",
+        time:       "Summer 2024",
+        position:   "Software Engineering Intern",
+        description: [
+            "TBD"
+        ]
+    },
+    {
+        logo:       "https://i.imgur.com/GGHOhcp.jpg",
+        name:       "VCI Construction, LLC",
+        shortname:  "VCI",
+        time:       "2021-2024",
+        position:   "IT Data Specialist",
+        description: [
+            "Database Management",
+            "Visual Basic Scripting",
+            "Microsoft Excel",
+            "Microsoft Power BI",
+            "Excel Automation",
+            "Refactoring",
+            "QA / QC",
+            "PM, CX, 3GIS Teamwork",
+            "Reports for Upper Management",
+        ]
+    },
+    {
+        logo:       "https://i.imgur.com/hqiI1Qy.png",
+        name:       "Paramount+",
+        shortname:  "Paramount",
+        time:       "Summer 2023",
+        position:   "Software Engineering Intern",
+        description: [
+            "Agile",
+            "API",
+            "BeautifulSoup",
+            "C#",
+            "Git/Github",
+            "Jira",
+            "MySQL",
+            "Python",
+            "Sprints",
+            "Stand Ups",
+            "Visual Studio",
+        ]
+    },
+    {
         logo:       "https://i.imgur.com/ndv0uLl.jpg",
         name:       "Revel Energy, LLC",
         shortname:  "Revel",
@@ -18,51 +66,7 @@ export const workexperience = [
             "Proposal Development"
         ]
     },
-    {
-        logo:       "https://i.imgur.com/GGHOhcp.jpg",
-        name:       "VCI Construction, LLC",
-        shortname:  "VCI",
-        time:       "2021-2024",
-        position:   "Project Data Analyst",
-        description: [
-            "Database Management",
-            "Visual Basic Scripting",
-            "Microsoft Excel",
-            "Automation in Excel",
-            "Microsoft Power BI",
-            "Refactoring Code",
-            "QA / QC",
-            "PM, CX, 3GIS Teamwork",
-            "Reports for Upper Management",
-        ]
-    },
-    {
-        logo:       "https://i.imgur.com/hqiI1Qy.png",
-        name:       "Paramount+",
-        shortname:  "Paramount",
-        time:       "Summer 2023",
-        position:   "Software Engineering Intern",
-        description: [
-            "Agile",
-            "BeautifulSoup",
-            "C#",
-            "Git/Github",
-            "Jira",
-            "MySQL",
-            "Python",
-            "Sprints",
-            "Stand Ups",
-            "Visual Studio",
-        ]
-    },
-    {
-        logo:       "https://i.imgur.com/hqiI1Qy.png",
-        name:       "Paramount+",
-        shortname:  "Paramount",
-        time:       "Summer 2024",
-        position:   "Software Engineering Intern",
-        description: [
-            "TBD"
-        ]
-    },
+    
+    
+    
 ]
